requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
google-generativeai>=0.3.0

